pyproject.toml
[project]
name = "genesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.13",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "openai>=1.93.0",
    "hydra-core>=1.3.0",
]
